Started working on Restaurant Page project
Initially the website template was set. We added few javascript files to configure and check if the webpack is working. Then we created the actual website in index.html. After that we moved that part to javascript files so that the html file will be generated by our javascript files once we run the webpack command. Next we need to work on the tabbed window feature of this project.