Completed the tabbed UI part of Restaurant Page
Tabs were added and content was getting rendered on the go based on which tab is selected.
Each tab is a seperate javascript file which helps loading the content when that tab is selected.
Stylesheet was updated to make things look proper. lastly, this is optional but still we need to add css and images using webpack. We will be working on that in next part.